Implemented nested inline And Add QUiz Attempting

+++ Admin/views.py
@@ -1,11 +1,12 @@
+from django.forms import inlineformset_factory
 from django.shortcuts import render, redirect
 from Student.models import Student, Admin
-from .models import Question, Quiz, AnswerOptions
+from .models import Question, Quiz
 from django.contrib.auth import login
 from django.views import View
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
-from .forms import QuestionForm, QuizForm, QuestionFormSet, AnswerOptionForm, AnswerOptionFormSet
+from .forms import *
 from django.views.generic import TemplateView
 from django.shortcuts import get_object_or_404
 
@@ -34,13 +35,15 @@ class AdminLogin(View):
         return render(request, 'Admin/AdminLogin.html', {"error": error})
 
 
-@login_required()
 def home(request):
-    return render(request, 'Admin/Home.html')
+    if request.user.is_authenticated and request.user.is_admin:
+        return render(request, 'Admin/Home.html')
+    else:
+        return redirect('adminhome')
 
 
 def Students(request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_admin:
         students = Student.objects.all()
         return render(request, 'Admin/Students.html', {'students': students})
     else:
@@ -48,7 +51,7 @@ def Students(request):
 
 
 def Quizes(request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_admin:
         quizes = Quiz.objects.all()
         return render(request, 'Admin/Quizes.html', {'quizes': quizes})
     else:
@@ -57,16 +60,21 @@ def Quizes(request):
 
 class Quiz_detail(View):
     def get(self, request, pk):
-        if request.user.is_authenticated:
-            quizobj = get_object_or_404(Quiz, pk=pk)
+        if request.user.is_authenticated and request.user.is_admin:
+            quizobj = Quiz.objects.get(pk=pk)
+            print(quizobj.assigned_to)
             questions = Question.objects.filter(quiz=pk)
-
-            answer_options = AnswerOptions.objects.filter(question__quiz=pk)
-
+            question_options = {}
+            for question in questions:
+                print("Question id:", question.id)
+                options = AnswerOptions.objects.filter(question=question)
+                option_texts = [option.option_text for option in options]
+                question_options[question.id] = option_texts
+            print(question_options)
             return render(request, 'Admin/QuizDetails.html', {
                 'quiz': quizobj,
                 'questions': questions,
-                'answer_options': answer_options
+                'question_options': question_options,
             })
         else:
             return redirect('adminhome')
@@ -77,50 +85,59 @@ class Quiz_detail(View):
 
 class CreateQuiz(View):
     def get(self, request):
-        if request.user.is_authenticated:
+        if request.user.is_authenticated and request.user.is_admin:
             quiz_form = QuizForm()
             question_formset = QuestionFormSet(prefix='questions')
-            return render(request, 'Admin/CreateQuiz.html', {
+            options_formset = AnswerOptionFormSet(prefix='options')
+            return render(request, 'Admin/createQuiz.html', {
                 'quizform': quiz_form,
-                'question_formset': question_formset
+                'question_formset': question_formset,
+                'options_formset': options_formset
             })
         else:
             return redirect('adminhome')
 
     def post(self, request):
+        # print(request.POST)
         quiz_form = QuizForm(request.POST)
         if quiz_form.is_valid():
-            print("Quiz form is valid")
             quiz = quiz_form.save(commit=False)
-
             question_formset = QuestionFormSet(request.POST, instance=quiz, prefix='questions')
             if question_formset.is_valid():
-                print("Question formset is valid")
                 quiz.save()
+
+                assigned_to_ids = request.POST.getlist('assigned_to')
+                assigned_to_students = Student.objects.filter(pk__in=assigned_to_ids)
+                quiz.assigned_to.set(assigned_to_students)
+
                 questions = question_formset.save(commit=False)
                 for question in questions:
                     question.quiz = quiz
                     question.save()
 
+                    options_formset = AnswerOptionFormSet(request.POST, prefix='options', instance=question)
+                    if options_formset.is_valid():
+                        options = options_formset.save(commit=False)
+                        for option in options:
+                            option.question = question
+                            option.save()
+
                 return redirect('quizes')
 
             else:
-                print("Question formset is not valid")
-                return render(request, 'Admin/CreateQuiz.html', {
+                return render(request, 'Admin/createQuiz.html', {
                     'quizform': quiz_form,
-                    'formset': question_formset,
+                    'question_formset': question_formset,
                 })
         else:
-            print("Quiz form is not valid")
-            print(quiz_form.errors.as_data())
-            return render(request, 'Admin/CreateQuiz.html', {
+            return render(request, 'Admin/createQuiz.html', {
                 'quizform': quiz_form
             })
 
 
 class CreateQuestion(View):
     def get(self, request):
-        if request.user.is_authenticated:
+        if request.user.is_authenticated and request.user.is_admin:
             question_form = QuestionForm()
             formset = AnswerOptionFormSet()
             return render(request, 'Admin/CreateQuestion.html', {'question_form': question_form, 'formset': formset})
@@ -170,7 +187,7 @@ class CreateQuestion(View):
 
 class Quiz_Delete(View):
     def post(self, request, pk):
-        if request.user.is_authenticated:
+        if request.user.is_authenticated and request.user.is_admin:
             quiz = get_object_or_404(Quiz, pk=pk)
             quiz.delete()
             return redirect('quizes')
@@ -184,7 +201,7 @@ class AddOption(TemplateView):
     template_name = "Admin/CreateQuestion.html"
 
     def get(self, *args, **kwargs):
-        formset = AnswerOptionFormSet(queryset=AnswerOptions.objects.none())
+        formset = AnswerOptionFormSet()
         return self.render_to_response({'formset': formset})
 
     def post(self, *args, **kwargs):
